+++ src/utils/debounce.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export default function (debouncedFn: (...args: any[]) => void, delay: number) {
+export default function (debouncedFn: (...args: any[]) => void, delay = 500) {
   let timeout: NodeJS.Timeout | null = null
 
   return (...params: unknown[]) => {